+++ front/src/components/CalendarBoard/index.jsx
@@ -1,24 +1,9 @@
 import React from "react";
 import { GridList, StylesProvider } from "@material-ui/core";
 import * as styles from "./style.css";
-import dayjs from "dayjs";
-import "dayjs/locale/ja";
 
-dayjs.locale("ja");
-
-const createCalendar = () => {
-  const firstDay = dayjs().startOf("month");
-  const firstDayIndex = firstDay.day();
-
-  return Array(35)
-    .fill(0)
-    .map((_, i) => {
-      const diffFromFirstDay = i - firstDayIndex;
-      const day = firstDay.add(diffFromFirstDay, "day");
-
-      return day;
-    });
-}
+import CalendarElement from "../CalendarElement";
+import { createCalendar } from "../../services/calendar";
 
 const calendar = createCalendar();
 
@@ -28,7 +13,7 @@ const CalendarBoard = () => {
       <GridList className={styles.grid} cols={7} spacing={0} cellHeight="auto">
         {calendar.map(c => (
           <li key={c.toISOString()}>
-            <div className={styles.element}>{c.format("D")}</div>
+            <CalendarElement>{c.format("D")}</CalendarElement>
           </li>
         ))}
       </GridList>
