+++ front/src/components/CalendarElement/style.css
@@ -1,0 +1,5 @@
+.element {
+  border-right: 1px solid #ccc;
+  border-bottom: 1px solid #ccc;
+  height: 18vh;
+}
