+++ front/src/index.jsx
@@ -1,6 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import CalendarBoard from "./components/CalendarBoard";
+import dayjs from "dayjs";
+import "dayjs/locale/ja";
+dayjs.locale("ja");
 
 const App = () => (
   <div>
